requirements.txt
streamlit>=1.30.0
openai-whisper
sounddevice>=0.4.6
soundfile>=0.12.1
numpy
scipy
python-dotenv>=1.0.0
scikit-learn>=1.3.0
torch>=2.1.0
torchaudio>=2.1.0
pyannote.audio>=3.1.0